refactor: restructure layouts and update landing page content
